+++ src/main/kotlin/io/kjson/uri/FragmentElement.kt
@@ -28,11 +28,17 @@ package io.kjson.uri
 import io.kjson.uri.Element.Companion.encodeReserved
 import net.pwall.text.UTF8StringMapper.encodeUTF8
 
-class FragmentElement(val name: String) : Element {
+class FragmentElement(val names: List<String>) : Element {
 
     override fun appendTo(a: Appendable, variables: List<Variable>) {
-        a.append('#')
-        variables.find { it.name == name }?.value?.let { a.append(it.toString().encodeUTF8().encodeReserved()) }
+        var continuation = false
+        for (name in names) {
+            variables.find { it.name == name }?.value?.let {
+                a.append(if (continuation) ',' else '#')
+                a.append(it.toString().encodeUTF8().encodeReserved())
+                continuation = true
+            }
+        }
     }
 
 }